Migration to SDL3 + constructor modifications

+++ version2/Source/Separator.cpp
@@ -14,7 +14,7 @@ void Separator::handle_event(SDL_Event& event) {}
 
 
 bool Separator::is_mouse_in() {
-	int x, y;
+	float x, y;
 	SDL_GetMouseState(&x, &y);
 	return ((x > xBegin_) && (x < xEnd_));
 }
@@ -22,9 +22,9 @@ bool Separator::is_mouse_in() {
 
 Separator::Separator() {
 	int i = 0;
-	while(i*TILE_DIM < WINDOW_HEIGHT) {
-		Bush *bushLeft = new Bush({xBegin_, i*TILE_DIM}),
-				 *bushRight = new Bush({xEnd_ - TILE_DIM, i*TILE_DIM});
+	while((int)i*TILE_DIM < WINDOW_HEIGHT) {
+		Bush *bushLeft = new Bush({xBegin_, (float)i*TILE_DIM}),
+			 *bushRight = new Bush({xEnd_ - TILE_DIM, (float)i*TILE_DIM});
 		vTextures.push_back(bushLeft);
 		vTextures.push_back(bushRight);
 		i++;
